View and Search Users
Used the template in the account holders dictionary to create the view users command. It now displays all users and their debit transactions history.

I'm still working on the search users command, I cant seem to isolate user info and print it in a visually attractive way. It is currently incomplete.